+++ tests/test_container.py
@@ -14,6 +14,7 @@ from dantro.base import BaseDataContainer, CheckDataMixin
 from dantro.base import ItemAccessMixin
 from dantro.mixins.base import UnexpectedTypeWarning
 from dantro.mixins.proxy_support import Hdf5ProxySupportMixin
+from dantro.groups import OrderedDataGroup
 from dantro.containers import MutableSequenceContainer
 from dantro.containers import NumpyDataContainer, XrDataContainer
 from dantro.proxy import Hdf5DataProxy
@@ -487,11 +488,11 @@ def test_XrDataContainer():
 
 
     # linked mapping . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
-    with pytest.raises(NotImplementedError):
-        XrDataContainer(name="xrdc", data=np.arange(10),
-                        attrs=dict(dims=['time'],
-                                   coords__time="../foo",
-                                   coords_mode__time='linked'))
+    xrdc = XrDataContainer(name="xrdc", data=np.arange(10),
+                           attrs=dict(dims=['time'],
+                                      coords__time="",  # refer to itself
+                                      coords_mode__time='linked'))
+    # NOTE This works properly only when using proxies; tested seperately
 
     # Invalid coordinate mode . . . . . . . . . . . . . . . . . . . . . . . . .
     with pytest.raises(ValueError, match="Invalid mode 'invalid' to interpre"):
@@ -550,7 +551,7 @@ def test_XrDataContainer_proxy_support(tmp_h5_dset):
     attrs = dict(foo="bar", dims=['x', 'y', 'z'],
                  coords__x=['1 m'], coords__z=['1 cm', '2 cm', '3 cm'])
     
-    # Check that proxy support isenabled now
+    # Check that proxy support is enabled
     assert Hdf5ProxyXrDC.DATA_ALLOW_PROXY
 
     # Create a proxy
@@ -775,3 +776,65 @@ def test_XrDataContainer_dask_integration(tmp_h5file):
 
     assert xrdc.data_is_proxy
     assert xrdc_nodask.data_is_proxy
+
+def test_XrDataContainer_linked_coordinates(tmp_h5_dset):
+    """Test 'linked' and 'from_path' coordinate modes for XrDataContainer"""
+
+    class Hdf5ProxyXrDC(Hdf5ProxySupportMixin, XrDataContainer):
+        pass
+    
+    # Check that proxy support is enabled
+    assert Hdf5ProxyXrDC.DATA_ALLOW_PROXY
+
+    # Create a proxy
+    proxy = Hdf5DataProxy(obj=tmp_h5_dset)  # shape: (1,2,3)
+
+    # Create a XrDataContainer with proxy support and linked coordinates
+    xrdc = Hdf5ProxyXrDC(name="xrdc", data=proxy,
+                         attrs=dict(dims=['x', 'y', 'z'],
+                                    coords_mode__x='linked',
+                                    coords__x='../some_other_data',
+                                    coords_mode__y='linked',
+                                    coords__y='../../coords/y',
+                                    coords_mode__z='linked',
+                                    coords__z='../../coords/more/z'))
+    
+    # Should have succeeded and be a proxy now
+    assert xrdc.data_is_proxy
+
+    # Now, incorporate it into a tree
+    root = OrderedDataGroup(name="root")
+
+    g_data = root.new_group("data")
+    g_data.add(xrdc)
+    g_data.new_container("some_other_data", Cls=XrDataContainer,
+                         data=[3.14])
+
+    g_coords = root.new_group("coords")
+    g_coords.new_container("y", Cls=XrDataContainer, data=[23, 42])
+
+    g_more_coords = g_coords.new_group("more")
+    g_more_coords.new_container("z", Cls=XrDataContainer, data=[2, 4, 8])
+
+    # The original data should still be proxy
+    assert xrdc.data_is_proxy
+
+    # Now, resolving it should lead to link resolution ... which happens in the
+    # backround and just leads to the data being available as coordiantes
+    assert (xrdc.coords['x'] == [3.14]).all()
+    assert (xrdc.coords['y'] == [23, 42]).all()
+    assert (xrdc.coords['z'] == [2, 4, 8]).all()
+
+
+    # Link resolution should fail if not embedded in a data tree
+    lone_xrdc = Hdf5ProxyXrDC(name="xrdc", data=proxy,
+                              attrs=dict(dims=['x', 'y', 'z'],
+                                         coords_mode__x='linked',
+                                         coords__x='../some_other_data',
+                                         coords_mode__y='linked',
+                                         coords__y='../../coords/y',
+                                         coords_mode__z='linked',
+                                         coords__z='../../coords/more/z'))
+
+    with pytest.raises(RuntimeError, match="Failed resolving target of link"):
+        lone_xrdc.coords['x']